codelift exports inspect & register
