+++ packages/codelift/env.js
@@ -1,0 +1,2 @@
+// An object to share values between register & inspect
+module.exports = {};
